Acquisition events: add getSource(); add AcquisitionSequenceStartedEvent.
This allows us to distinguish between acquisition events generated by the main
program and those generated by third-party code, and also doesn't tie the base
event class to the SequenceSettings object, which will eventually be
deprecated.
