+++ mmstudio/src/main/java/org/micromanager/events/AcquisitionSequenceStartedEvent.java
@@ -18,36 +18,22 @@
 //               CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
 //               INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES.
 
-package org.micromanager.events.internal;
+package org.micromanager.events;
 
 import org.micromanager.data.Datastore;
-import org.micromanager.events.AcquisitionSequenceStartedEvent;
 import org.micromanager.SequenceSettings;
 
-public class DefaultAcquisitionStartedEvent implements AcquisitionSequenceStartedEvent {
-   private Datastore store_;
-   private Object source_;
-   private SequenceSettings settings_;
-
-   public DefaultAcquisitionStartedEvent(Datastore store, Object source,
-         SequenceSettings settings) {
-      store_ = store;
-      source_ = source;
-      settings_ = settings;
-   }
-
-   @Override
-   public Datastore getDatastore() {
-      return store_;
-   }
-
-   @Override
-   public Object getSource() {
-      return source_;
-   }
-
-   @Override
-   public SequenceSettings getSettings() {
-      return settings_;
-   }
+/**
+ * This variant of the AcquisitionStartedEvent is used for acquisitions that
+ * can be described by a SequenceSettings; it provides access to those
+ * SequenceSettings.
+ */
+public interface AcquisitionSequenceStartedEvent extends AcquisitionStartedEvent {
+   /**
+    * Return the SequenceSettings used to control the parameters of the
+    * acquisition. Note that the images in the datastore may not necessarily
+    * match these parameters (with respect to number of Z-slices, etc.) due
+    * to the actions of image processors in the data processing pipeline.
+    */
+   public SequenceSettings getSettings();
 }
