+++ mmstudio/src/main/java/org/micromanager/events/AcquisitionStartedEvent.java
@@ -25,8 +25,9 @@ import org.micromanager.SequenceSettings;
 
 /**
  * This class signals that an acquisition is starting, and provides access to
- * the Datastore that images from the acquisition will be put into, as well
- * as the parameters governing how the acquisition will run.
+ * the Datastore that images from the acquisition will be put into. Third-party
+ * code may subclass this event to provide notifications of their own
+ * acquisitions.
  */
 public interface AcquisitionStartedEvent {
    /**
@@ -36,10 +37,11 @@ public interface AcquisitionStartedEvent {
    public Datastore getDatastore();
 
    /**
-    * Return the SequenceSettings used to control the parameters of the
-    * acquisition. Note that the images in the datastore may not necessarily
-    * match these parameters (with respect to number of Z-slices, etc.) due
-    * to the actions of image processors in the data processing pipeline.
+    * Return an Object used to identify the entity in charge of the
+    * acquisition. This can be used by recipients to distinguish different
+    * types of acquisitions. If you have code that publishes
+    * AcquisitionStartedEvents, then that code should re-use the same object
+    * for each event.
     */
-   public SequenceSettings getSettings();
+   public Object getSource();
 }
